Star Transit: Comment Tm typetapes related classes for time being. Refactored existing filetype to parse also the tms.
